add limit to base-query

+++ module/core/src/test/java/ru/protei/portal/test/service/ContactServiceTest.java
@@ -48,7 +48,7 @@ public class ContactServiceTest {
 
         Assert.assertNotNull(service);
 
-        CoreResponse<List<Person>> result = service.contactList(new ContactQuery(null, "Михаил", En_SortField.person_full_name, En_SortDir.ASC));
+        CoreResponse<List<Person>> result = service.contactList(new ContactQuery((Long)null, "Михаил", En_SortField.person_full_name, En_SortDir.ASC));
 
         Assert.assertNotNull(result);
         Assert.assertTrue(result.isOk());
